Use invokedynamic to call methods (+ minor fixes)
Our method calls are now polymorphic in the receiver, like they should
be, and generic otherwise. Java's type system is not used anymore: all
user-facing Core methods take and return BObjects.

The fact that we use invokedynamic means that the JVM might (no
guarantees here) be able to notice you're dynamically invoking the same
method all the time and inline the dynamic machinery to make it as fast
as a statically compiled call.

Also comes with a few minor improvements:
- update pom.xml to have ANTLR also generate visitors
- clean up BInt methods a bit
- add comparison ops for BInt