Special-case test* methods to generate trampolines
This commit introduces a few features to make testing easier. It
special-cases methods that start with 'test' and take no arguments, then
generates a 'trampoline method'. This trampoline method is simply the
unmangled name of the method, and it does nothing more than calling the
mangled version. This should make it easier to test Babble code from
external test runners.

It also restructures SelfTest to match the convention.

For now, a var reference of 'this' will point to a new instance of the
class currently being generated. This should really be the existing
instance. A TODO has been added.