Add very basic support for blocks
Every Block (closure block) is compiled to its own class. Each class is
a subclass of BBlock and has a value method with the block's contents as
its body.

Classes are named after the enclosing class with a dollar sign and a
number attached, like Java inner classes.

There is no support for empty blocks, or blocks with arguments, and the
fact that return works depends on the fact that we're not cleaning up
the stack as much as we should, but it's a start.