Introduce Prelude, move bits of runtime to Babble
This commit introduces the Prelude file, which has Babble classes that extend
`org.twnc.runtime` classes. That way, part of the desired behaviour can be
expressed in Babble instead of in Java. It is the first step in getting rid of
most of the Java helper classes altogether.

At the moment, there is no support for arguments and no support for references
to `this`, so the amount of behaviour that can be ported is rather small.
Still, it's a start.