Pop all expressions except the last in sequences
The result of a sequence expression is the result of its last part. All the
other parts are ignored. In the current system, they accumulated on the stack
for no good reason. This change inserts pop instructions to clean them up and
reduce the stack space required for each method.

It also removes the BNil stack guard at the bottom of the stack: in our case,
it would be better to have the JVM verifier warn us of a stack problem than
silently return Nil.