Add first version of report
This adds a first version of the report, in LaTeX, with some empty sections and
the beginnings of a 'language summary' section.

It also adds a slightly cleaned up, syntax-highlighted version of the ANTLR
grammar as well as a syntax-highlighted ASTGenerator to the report. The
syntax-highlighted versions have been bundled so that the report compiles
correctly even if pygments is not installed.