Add metaclasses, respondsTo, introspection pass
All instances of Babble classes (not runtime classes yet) can now be
queried for their class at runtime. This allows for a very basic form of
introspection:

    (true class name == "True") assert.
    (true class class name == "Class") assert.

Instances of BClass can also tell you if an object will respond to a
certain selector:

    (this class respondsTo:"main") assert.
    (this class respondsTo:"bogusMethod") not assert.
    (this class class repondsTo:"name") assert.

(Note that respondsTo is a method on the class in Babble. It is a method
on the object in Smalltalk.)