Replace 'special' variables with globals
This replaces the isSpecial method on ScopeStack with a ProgramNode-wide
map of globals. These globals are true, false, nil, and defined class
names; the latter point to the metaclasses. For example, the following:

    bob := Duck new.

will make a new DuckClass metaclass instance, call `new` on that, and
assign the resulting Duck to `bob`.

Binding these globals is a last-resort and has been deferred for as long
as possible: we use them in the BytecodeGenerator, only if there is no
other possible way. This means you can 'redefine' the globals if you
want to, and it means that adding a class named Foo will not disturb
some field in some class that also happens to be called Foo.

Also moves Assignment in the grammar to fix a precedence issue where
`bob := Duck new` was parsed as `(bob := Duck) new`.